+++ lib/src/model/response/integer_response.g.dart
@@ -9,9 +9,11 @@ part of 'integer_response.dart';
 IntegerResponse _$IntegerResponseFromJson(Map<String, dynamic> json) =>
     IntegerResponse(
       sceneItemId: json['sceneItemId'] as int?,
+      sceneItemIndex: json['sceneItemIndex'] as int?,
     );
 
 Map<String, dynamic> _$IntegerResponseToJson(IntegerResponse instance) =>
     <String, dynamic>{
       'sceneItemId': instance.sceneItemId,
+      'sceneItemIndex': instance.sceneItemIndex,
     };