+++ lib/src/request/general.dart
@@ -51,14 +51,14 @@ class General {
   // TODO:
   Future<void> callVendorRequest() async => throw UnimplementedError();
 
-  ///Gets an array of all hotkey names in OBS
+  /// Gets an array of all hotkey names in OBS
   ///
   /// - Complexity Rating: 3/5
   /// - Latest Supported RPC Version: 1
   /// - Added in v5.0.0
   Future<List<String>> get hotkeyList async => await getHotkeyList();
 
-  ///Gets an array of all hotkey names in OBS
+  /// Gets an array of all hotkey names in OBS
   ///
   /// - Complexity Rating: 3/5
   /// - Latest Supported RPC Version: 1
@@ -76,9 +76,35 @@ class General {
     ));
   }
 
-  // TODO:
-  Future<void> triggerHotkeyByKeySequence() async => throw UnimplementedError();
+  /// Triggers a hotkey using a sequence of keys.
+  ///
+  /// - Complexity Rating: 4/5
+  /// - Latest Supported RPC Version: 1
+  /// - Added in v5.0.0
+  Future<void> triggerHotkeyByKeySequence(
+          {String? keyId, KeyModifiers? keyModifiers}) async =>
+      await obsWebSocket.sendRequest(Request(
+        'TriggerHotkeyByKeySequence',
+        requestData: {
+          'keyId': keyId,
+          'keyModifiers': keyModifiers?.toJson(),
+        }..removeWhere((key, value) => value == null),
+      ));
 
-  // TODO:
-  Future<void> sleep() async => throw UnimplementedError();
+  /// Sleeps for a time duration or number of frames. Only available in request batches with types SERIAL_REALTIME or SERIAL_FRAME.
+  ///
+  /// - Complexity Rating: 2/5
+  /// - Latest Supported RPC Version: 1
+  /// - Added in v5.0.0
+  Future<void> sleep({
+    required int sleepMillis,
+    required int sleepFrames,
+  }) async =>
+      await obsWebSocket.sendRequest(Request(
+        'Sleep',
+        requestData: {
+          'sleepMillis': sleepMillis,
+          'sleepFrames': sleepFrames,
+        },
+      ));
 }