+++ tool/grind.dart
@@ -1,6 +1,5 @@
 import 'package:grinder/grinder.dart';
 import 'package:mustache_template/mustache.dart';
-import 'package:process_run/which.dart';
 import 'package:obs_websocket/src/util/meta_update.dart';
 import 'package:pub_semver/pub_semver.dart';
 import 'package:pubspec/pubspec.dart';
@@ -16,12 +15,12 @@ main(args) => grind(args);
 @DefaultTask('Build the project.')
 @Depends(test)
 build() {
-  log("building...");
+  // log('building...');
 }
 
 @Task('publish')
-@Depends(analyze, version, doc, commit, dryrun)
-// @Depends(analyze, version, test, doc, commit, dryrun)
+@Depends(analyze, version, doc, commit, release, dryrun)
+// @Depends(analyze, version, test, doc, commit, release, dryrun)
 publish() {
   log('''
 Use the command:
@@ -32,8 +31,8 @@ To publish this package on the pub.dev site.
 }
 
 @Task('dart pub publish --dry-run')
-dryrun() {
-  shell(args: ['pub', 'publish', '--dry-run']);
+dryrun() async {
+  await shell(args: ['pub', 'publish', '--dry-run']);
 }
 
 @Task('dartdoc')
@@ -42,17 +41,13 @@ doc() {
   DartDoc.doc();
 }
 
-@Task('dart analyze')
+@Task('analyze')
 analyze() {
   Analyzer.analyze('.', fatalWarnings: true);
 }
 
 @Task('version')
 version() async {
-  final metaUpdate = MetaUpdate('pubspec.yaml');
-
-  metaUpdate.writeMetaDartFile('lib/src/util/meta.dart');
-
   final version = config['version']!;
 
   final newTag = await isNewTag(version);
@@ -61,27 +56,66 @@ version() async {
     updateMarkdown(config);
 
     await updatePubspec(version);
+
+    //uses the pubspec file so this needs to run after the pubspec update
+    MetaUpdate('pubspec.yaml').writeMetaDartFile('lib/src/util/meta.dart');
   }
 }
 
-@Task('commit to git')
+@Task('release')
+release() async {
+  final result = await shell(
+    exec: 'gh',
+    args: ['release', 'view', 'v${config['version']}'],
+  );
+
+  if (result.stderr.contains('not found')) {
+    await shell(
+      exec: 'gh',
+      args: [
+        'release',
+        'create',
+        'v${config['version']}',
+        '--title',
+        'Release v${config['version']}',
+        '--notes',
+        '${config['change']}',
+        '--repo',
+        '${config['repo']}'
+      ],
+      verbose: true,
+    );
+  }
+}
+
+@Task('commit')
 commit() async {
   final newTag = await isNewTag(config['version']);
 
-  shell(exec: 'git', args: ['add', '.']);
-
-  shell(exec: 'git', args: ['commit', '-m', '\'${config['change']}\'']);
-
-  if (newTag) {
-    shell(exec: 'git', args: ['tag', 'v${config['version']}']);
-
-    shell(exec: 'git', args: ['push', '--tags']);
+  try {
+    await shell(
+      exec: 'git',
+      args: ['commit', '-a', '-m', '\'${config['change']}\''],
+      verbose: true,
+    );
+  } catch (exception) {
+    log('No files committed');
   }
 
-  shell(exec: 'git', args: ['push']);
+  if (newTag) {
+    await shell(exec: 'git', args: ['tag', 'v${config['version']}']);
+
+    await shell(exec: 'git', args: ['push', '--tags']);
+  }
+
+  await shell(
+    exec: 'git',
+    args: ['push'],
+    verbose: true,
+  );
 }
 
-@Task('run tests')
+@Task('test')
 @Depends(version)
 test() {
   TestRunner().test();
@@ -97,25 +131,29 @@ YamlMap getConfig() {
   return config;
 }
 
-String shell(
-    {String exec = 'dart',
-    List<String> args = const <String>[],
-    bool verbose = true}) {
-  final executable = whichSync(exec);
-
-  if (executable == null) throw Exception();
-
-  return run(executable, arguments: args, quiet: !verbose);
-}
-
 Future<bool> isNewTag(String version) async {
-  final result = shell(
+  final result = await shell(
     exec: 'git',
     args: ['tag', '-l', 'v$version'],
     verbose: false,
   );
 
-  return result.trim() != 'v$version';
+  return result.stdout.trim() != 'v$version';
+}
+
+Future<ProcessResult> shell(
+    {String exec = 'dart',
+    List<String> args = const <String>[],
+    bool verbose = true}) async {
+  final result = await Process.run(exec, args);
+
+  if (verbose) {
+    log('stderr:\n${result.stderr}');
+
+    log('stdout:\n${result.stdout}');
+  }
+
+  return result;
 }
 
 void updateMarkdown(config) {